Student Attendance System/requirements.txt
numpy
pandas
Flask
Werkzeug
opencv-python
Pillow